+++ lib/chatgpt_flutter.dart
@@ -1,3 +1,6 @@
 library chatgpt_flutter;
 
-class OpenAI {}
+export 'src/client/exceptions/exceptions.dart';
+export 'src/client/open_ai_client.dart';
+export 'src/models/models.dart';
+export 'src/open_ai.dart';