+++ lib/src/client/exceptions/exceptions.dart
@@ -1,0 +1,2 @@
+export 'missing_key_exception.dart';
+export 'request_failed_exception.dart';