+++ lib/src/models/shared/configuration/configuration.dart
@@ -1,0 +1,2 @@
+export 'client_setup.dart';
+export 'open_ai_configuration.dart';